pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pycryptodome>=3.23.0",
    "scapy>=2.6.1",
    "streamlit>=1.50.0",
]
